Move 'immutable' ccompat helpers to shared ccompat package
