--- akka-actor/src/main/scala-2.13+/akka/util/ccompat/imm/package.scala
@@ -1,9 +1,0 @@
-/*
- * Copyright (C) 2018-2019 Lightbend Inc. <https://www.lightbend.com>
- */
-
-package akka.util.ccompat
-
-import scala.collection.immutable
-
-package object imm {}
