+++ akka-cluster-tools/src/test/scala/akka/cluster/singleton/ClusterSingletonLeavingSpeedSpec.scala
@@ -82,7 +82,7 @@ class ClusterSingletonLeavingSpeedSpec
 
     Cluster(from).join(Cluster(to).selfAddress)
     within(15.seconds) {
-      import akka.util.ccompat.imm._
+      import akka.util.ccompat._
       awaitAssert {
         Cluster(from).state.members.map(_.uniqueAddress) should contain(Cluster(from).selfUniqueAddress)
         Cluster(from).state.members.unsorted.map(_.status) should ===(Set(MemberStatus.Up))
