+++ akka-cluster/src/main/scala/akka/cluster/protobuf/ClusterMessageSerializer.scala
@@ -22,7 +22,7 @@ import akka.cluster.InternalClusterAction._
 import akka.cluster.routing.{ ClusterRouterPool, ClusterRouterPoolSettings }
 import akka.routing.Pool
 import akka.util.ccompat._
-import akka.util.ccompat.imm._
+import akka.util.ccompat._
 import com.typesafe.config.{ Config, ConfigFactory, ConfigRenderOptions }
 
 /**
